feat(enterprise): add as 5th InferenceMode card in AI Text Enhancement + Agent Mode
The previous commit added Enterprise as a third tab inside
ReasoningModelSelector — but upstream had already moved both AI Text
Enhancement and Agent Mode to a card-list InferenceModeSelector with 4
modes (OpenWhispr Cloud / Cloud Providers / Local / Self-Hosted), and
ReasoningModelSelector is now invoked with `mode="cloud"|"local"`
which hides its internal tabs entirely. So the earlier 3-way tab
addition was effectively dead code in the new UI — Enterprise was
unreachable from the Settings panel.

This commit reworks Enterprise to fit the InferenceModeSelector
pattern:

- Extend `InferenceMode` type with `"enterprise"`. Settings store
  validators accept it, and the migration helpers auto-detect it from
  saved `reasoningProvider` / `agentProvider` values
  (bedrock/azure/vertex → enterprise mode).
- Extract `EnterpriseSection` from `ReasoningModelSelector` into a
  standalone `src/components/EnterpriseSection.tsx`. Same content
  (Bedrock/Azure/Vertex sub-tabs + EnterpriseProviderConfig form),
  Azure and Vertex tabs stay disabled with the "Soon" badge until
  their runtime paths land.
- Strip the dead 3-way tabs and inline EnterpriseSection from
  `ReasoningModelSelector.tsx`. The component is back to its
  upstream-shaped Cloud/Local switcher (with the optional `mode` prop
  to hide the tabs when used as a sub-component).
- Add a 5th card to the `aiModes` array in `SettingsPage.AiModelsSection`
  with the Building2 icon, plus the matching toast key
  (`switchedEnterprise`). Render `<EnterpriseSection>` when
  `reasoningMode === "enterprise"`.
- Same wiring in `AgentModeSettings`: 5th card → `<EnterpriseSection>`
  block, with `agentProvider`/`agentModel` instead of the reasoning
  ones.
- i18n: new `enterprise` + `enterpriseDesc` keys under
  `settingsPage.aiModels.modes` and `agentMode.settings.modes`, plus
  `switchedEnterprise.{title,description}` toast under
  `settingsPage.aiModels.toasts`. Translated to all 9 non-English
  locales using each locale's existing card-label tone (de keeps
  loanword `Enterprise`, es `Empresarial`, fr/it loanword `Enterprise`
  matching the file's other tier-name conventions, ja `エンタープライズ`,
  pt `Corporativo`, ru `Корпоративный`, zh-CN `企业级`,
  zh-TW `企業級`).
