+++ src/locales/en/translation.json
@@ -1251,10 +1251,12 @@
         "providers": "Cloud Providers",
         "local": "Local",
         "selfHosted": "Self-Hosted",
+        "enterprise": "Enterprise",
         "openwhisprDesc": "Reliable accuracy. No setup needed.",
         "providersDesc": "Bring your own API key.",
         "localDesc": "On-device models. Fully private.",
-        "selfHostedDesc": "Your own server on your network."
+        "selfHostedDesc": "Your own server on your network.",
+        "enterpriseDesc": "Use your organization's cloud account (AWS, Azure, GCP)."
       },
       "openwhisprCloud": "OpenWhispr Cloud",
       "openwhisprCloudDescription": "Reliable accuracy. No setup needed.",
@@ -1279,6 +1281,10 @@
         "switchedSelfHosted": {
           "title": "Self-Hosted",
           "description": "Using your own server."
+        },
+        "switchedEnterprise": {
+          "title": "Enterprise",
+          "description": "Using your organization's cloud account."
         }
       }
     },
@@ -2173,10 +2179,12 @@
         "providers": "Cloud Providers",
         "local": "Local",
         "selfHosted": "Self-Hosted",
+        "enterprise": "Enterprise",
         "openwhisprDesc": "Reliable accuracy. No setup needed.",
         "providersDesc": "Bring your own API key.",
         "localDesc": "On-device models. Fully private.",
-        "selfHostedDesc": "Your own server on your network."
+        "selfHostedDesc": "Your own server on your network.",
+        "enterpriseDesc": "Use your organization's cloud account (AWS, Azure, GCP)."
       },
       "openwhisprCloud": "OpenWhispr Cloud",
       "openwhisprCloudDescription": "Reliable accuracy. No setup needed.",
